# services/api-gateway/src/main/java/com/apipratudo/gateway/scheduling/dto/CustomerRequest.java
package com.apipratudo.gateway.scheduling.dto;

import jakarta.validation.constraints.NotBlank;

public class CustomerRequest {

  @NotBlank
  private String name;

  private String phone;

  private String email;

  public String getName() {
    return name;
  }

  public void setName(String name) {
    this.name = name;
  }

  public String getPhone() {
    return phone;
  }

  public void setPhone(String phone) {
    this.phone = phone;
  }

  public String getEmail() {
    return email;
  }

  public void setEmail(String email) {
    this.email = email;
  }
}